sonic-o1/evaluation_inference/external_repos/Uni-MoE/Uni-MoE-2/requirement.txt
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.10.0
apted==1.0.3
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==15.0.0
bitarray==3.7.1
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
clean-fid==0.1.35
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
Cython==3.1.3
DataProperty==1.1.0
datasets==2.20.0
debugpy==1.8.17
decorator==5.2.1
decord==0.6.0
deepspeed==0.15.4
diffusers==0.30.2
dill==0.3.8
Distance==0.1.3
distlib==0.4.0
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
editdistance==0.8.1
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.5
executing==2.2.0
fairscale==0.4.13
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.5.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.14.0
h5py==3.12.1
hf-xet==1.1.5
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.34.2
identify==2.6.13
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.1
importlib_metadata==8.7.0
ipdb==0.13.13
ipykernel==7.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.9
langdetect==1.0.9
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
modelscope==1.27.0
more-itertools==10.7.0
moviepy==2.2.1
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numexpr==2.11.0
numpy==1.26.4
openai-whisper==20250625
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.12.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
polygon==1.2.7
polygon-api-client==1.15.3
Polygon3==3.0.9.1
pooch==1.8.2
portalocker==2.0.0
pre_commit==4.3.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pybind11==3.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
sacrebleu==1.5.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
sentry-sdk==2.33.2
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.3.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.21.4
torchdiffeq==0.2.5
tornado==6.5.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
typepy==1.3.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zss==1.2.0
zstandard==0.24.0
